CDPD-34177 HIVE-25306: Change Date/Timestamp parser from LENIENT to STRICT (Ashish Kumar Sharma, reviewed by Adesh Kumar Rao, Matt McCline, Sankar Hariappan, Stamatis Zampetakis)
With this change parsing string literals to dates/timestamps does not tolerate
dates which do not conform exactly to the expected syntax and does not compensate
for months, days, etc., out of the usual ranges.

When the literal cannot be parsed the null value is returned adopting
the default MySQL behavior.

This is a breaking change.

Closes #2445

(cherry picked from commit ba96607e31c6ee6dcf2a7a60ac3f954d960d95ba)

Change-Id: I0e8e35dccc943162540044edf993a25dfc0743b5