CDPD-42578: HIVE-25528: Avoid recalculating types after CBO on second AST pass (#2722) (Stephen Carlin reviewed by Alessandro Solimando, Zoltan Haindrich)
At compile time, Hive parses the query into ASTNodes.  For CBO, the
ASTNodes get converted into Calcite RexNodes and then get converted
back into ASTNodes. After the optimization step is done, the types
chosen for each step should be the final type. This commit eliminates
any type changes done on the conversion back to the ASTNode.

The reason for the ptest file changes are because there is an extra
constant fold done on the conversion back to the ASTNode. This
constant fold should have been done at optimization time. A Jira
will be filed for this, but there is no adverse effect for this issue.

Backport notes: The upstream HIVE-25528 patch was created on the Hive
code that uses a later version of Calcite (1.21) whereas here we have
to work with Calcite 1.19. Hence the fix has been modified accordingly
but functionally it provides the same behavior.

Change-Id: I82c74d876af48cd94934977f257e29d9e9763f9a