CDPD-65048: HIVE-26208: Exception in Vectorization with Decimal64 to Decimal casting (#3344)
The exception was happening because the code that casts the Decimal64 to
Decimal was not being added in filtered expressions.

Once this code was added, it caused a regression in check_constraint.q. The
reason for this was because we do not want to convert Decimal64 to Decimal
if the expression explicitly handles decimal64 types. A method was added
to these classes that will prevent the conversion in these cases.

(cherry picked from commit a694ee7b6da82a9d11312d32524bebd70d292a94)

Change-Id: I04237d95cd72634f727e6116f220189fb4832e97